Reverted back to 32 bit trace for x86 to keep flow working. This will be updated when the new PS kernel is built. Deprecated Config reader settings are removed.

+++ src/runtime_src/xdp/profile/plugin/aie_trace_new/x86/aie_trace.cpp
@@ -73,7 +73,7 @@ namespace xdp {
     auto tiles = metadata->getTilesForTracing();
 
     metadata->setTraceStartControl();
-    uint64_t delayCycles = metadata->getDelay();
+    uint32_t delayCycles = static_cast<uint32_t>(metadata->getDelay());
     bool userControl = metadata->getUseUserControl();//xrt_core::config::get_aie_trace_settings_start_type() == "kernel_event0";
     bool useDelay = metadata->getUseDelay();
